+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 [https://github.com/stickyio/link_stickyio_preview/commit/90f666d03cc8b18d227553c8ccbfa65570e97349](https://github.com/stickyio/link_stickyio_preview/commit/90f666d03cc8b18d227553c8ccbfa65570e97349)
 [https://github.com/stickyio/link_stickyio_preview/commit/b962dce3365c296a86d61f1f21a53086516ffea1](https://github.com/stickyio/link_stickyio_preview/commit/b962dce3365c296a86d61f1f21a53086516ffea1)
 [https://github.com/stickyio/link_stickyio_preview/commit/b9a731bae46e126905f881c662dc2c4c9fe61f56](https://github.com/stickyio/link_stickyio_preview/commit/b9a731bae46e126905f881c662dc2c4c9fe61f56)
+[https://github.com/stickyio/link_stickyio_preview/commit/90ac8c44586563268f093621bc3a67d3a9edde82](https://github.com/stickyio/link_stickyio_preview/commit/90ac8c44586563268f093621bc3a67d3a9edde82)
 ### Changed
 - Removed unused debug variables and fixed linting complaints
 - Fixed major bug in `getActiveBillingModels()` where it was not looping through available offer objects